Set up CMake configure dependencies on all JSON files used in the configuration
