+++ tools/python/mbed_tools/cli/main.py
@@ -15,7 +15,6 @@ from mbed_tools.lib.logging import set_log_level, MbedToolsHandler
 from mbed_tools.cli.configure import configure
 from mbed_tools.cli.list_connected_devices import list_connected_devices
 from mbed_tools.cli.project_management import new, import_, deploy
-from mbed_tools.cli.build import build
 from mbed_tools.cli.sterm import sterm
 from mbed_tools.cli.cmsis_mcu_descr import cmsis_mcu_descr
 
@@ -78,7 +77,6 @@ cli.add_command(list_connected_devices, "detect")
 cli.add_command(new, "new")
 cli.add_command(deploy, "deploy")
 cli.add_command(import_, "import")
-cli.add_command(build, "compile")
 cli.add_command(sterm, "sterm")
 cli.add_command(cmsis_mcu_descr)
