+++ tools/python/python_tests/mbed_tools/cli/test_configure.py
@@ -15,14 +15,14 @@ class TestConfigureCommand(TestCase):
     @mock.patch("mbed_tools.cli.configure.generate_config")
     @mock.patch("mbed_tools.cli.configure.MbedProgram")
     def test_generate_config_called_with_correct_arguments(self, program, generate_config):
-        CliRunner().invoke(configure, ["-m", "k64f", "-t", "gcc_arm"])
+        CliRunner().invoke(configure, ["-m", "k64f", "-t", "gcc_arm", "--cmake-build-dir", "build"])
 
         generate_config.assert_called_once_with("K64F", "GCC_ARM", program.from_existing())
 
     @mock.patch("mbed_tools.cli.configure.generate_config")
     @mock.patch("mbed_tools.cli.configure.MbedProgram")
     def test_generate_config_called_with_mbed_os_path(self, program, generate_config):
-        CliRunner().invoke(configure, ["-m", "k64f", "-t", "gcc_arm", "--mbed-os-path", "./extern/mbed-os"])
+        CliRunner().invoke(configure, ["-m", "k64f", "-t", "gcc_arm", "--mbed-os-path", "./extern/mbed-os", "--cmake-build-dir", "build"])
 
         generate_config.assert_called_once_with("K64F", "GCC_ARM", program.from_existing())
 
@@ -32,7 +32,7 @@ class TestConfigureCommand(TestCase):
         program = program.from_existing()
         custom_targets_json_path = pathlib.Path("custom", "custom_targets.json")
         CliRunner().invoke(
-            configure, ["-t", "gcc_arm", "-m", "k64f", "--custom-targets-json", custom_targets_json_path]
+            configure, ["-t", "gcc_arm", "-m", "k64f", "--custom-targets-json", custom_targets_json_path, "--cmake-build-dir", "build"]
         )
 
         generate_config.assert_called_once_with("K64F", "GCC_ARM", program)
@@ -43,7 +43,7 @@ class TestConfigureCommand(TestCase):
     def test_custom_output_directory_used_when_passed(self, program, generate_config):
         program = program.from_existing()
         output_dir = pathlib.Path("build")
-        CliRunner().invoke(configure, ["-t", "gcc_arm", "-m", "k64f", "-o", output_dir])
+        CliRunner().invoke(configure, ["-t", "gcc_arm", "-m", "k64f", "-o", output_dir, "--cmake-build-dir", "build"])
 
         generate_config.assert_called_once_with("K64F", "GCC_ARM", program)
         self.assertEqual(program.files.cmake_build_dir, output_dir)
@@ -54,7 +54,7 @@ class TestConfigureCommand(TestCase):
         program = program.from_existing()
         app_config_path = pathlib.Path("alternative_config.json")
         CliRunner().invoke(
-            configure, ["-t", "gcc_arm", "-m", "k64f", "--app-config", app_config_path]
+            configure, ["-t", "gcc_arm", "-m", "k64f", "--app-config", app_config_path, "--cmake-build-dir", "build"]
         )
 
         generate_config.assert_called_once_with("K64F", "GCC_ARM", program)
@@ -62,20 +62,19 @@ class TestConfigureCommand(TestCase):
 
     @mock.patch("mbed_tools.cli.configure.generate_config")
     @mock.patch("mbed_tools.cli.configure.MbedProgram")
-    def test_profile_used_when_passed(self, program, generate_config):
+    def test_cmake_build_dir_passed(self, program, generate_config):
         test_program = program.from_existing()
         program.reset_mock()  # clear call count from previous line
 
         toolchain = "gcc_arm"
         target = "k64f"
-        profile = "release"
 
         CliRunner().invoke(
-            configure, ["-t", toolchain, "-m", target, "--profile", profile]
+            configure, ["-t", toolchain, "-m", target, "--cmake-build-dir", "some-build-dir"]
         )
 
         program.from_existing.assert_called_once_with(
             pathlib.Path("."),
-            pathlib.Path(target.upper(), profile, toolchain.upper())
+            pathlib.Path("some-build-dir")
         )
         generate_config.assert_called_once_with("K64F", "GCC_ARM", test_program)
